// src/app/enums/paths.ts
export enum Paths {
    BASE = 'https://fierce-shore-48603.herokuapp.com/',
    PERSONA = 'api/v1/persona',
    CONTACT = 'api/v1/contact',
    TECHNOLOGY = 'api/v1/technology',
    JOB = 'api/v1/job',
    LOGIN = 'login',
    ABOUT = 'api/v1/about',
    EDUCATION = 'api/v1/education',
    PROJECT = 'api/v1/project'
}